+++ EDDI/MainWindow.xaml.cs
@@ -515,7 +515,9 @@ namespace Eddi
 
             if (!fromVA)
             {
+                SpeechService.Instance.Say(EDDI.Instance.Ship, "Goodbye.", true, 1);
                 EDDI.Instance.Stop();
+                Application.Current.ShutdownMode = ShutdownMode.OnExplicitShutdown;
                 Application.Current.Shutdown();
             }
         }