+++ SpeechResponder/HelpWindow.xaml.cs
@@ -1,18 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.Reflection;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
+using Utilities;
 
 namespace EddiSpeechResponder
 {
@@ -26,7 +16,17 @@ namespace EddiSpeechResponder
             InitializeComponent();
 
             // Read Markdown and convert it to HTML
-            string markdown = File.ReadAllText("Help.md");
+            string markdown;
+            try
+            {
+                DirectoryInfo dir = new DirectoryInfo(System.IO.Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location));
+                markdown = File.ReadAllText(dir.FullName + @"\Help.md");
+            }
+            catch (Exception ex)
+            {
+                Logging.Error("Failed to find variables.md", ex);
+                markdown = "";
+            }
             string html = CommonMark.CommonMarkConverter.Convert(markdown);
 
             // Insert the HTML