+++ ChangeLog.md
@@ -1,3 +1,10 @@
+#2.0.9
+  * Do not update ship configuration when data is not available from the companion API
+  * Be a little harsher when shutting down speech threads on close
+  * Make home station accessible to scripts
+  * Make current station more dynamic to match reality
+  * Fix crash when bringing up help windows from VoiceAttack
+
 #2.0.8
   * Update coriolis export with additional properties for improved accuracy
   * Handle "Profile unavailable" response from companion API